+++ src/main/java/dev/abarmin/aml/dashboard/BlockUpdateHandler.java
@@ -60,7 +60,7 @@ public class BlockUpdateHandler {
 
     if (!bindingResult.hasErrors()) {
       updateBlock(blockModel);
-      rebuildViewModel(model, blockModel);
+      rebuildViewModel(model, blockModel.getBlockId());
     }
     return "private/dashboard";
   }
@@ -68,8 +68,8 @@ public class BlockUpdateHandler {
   @PostMapping(params = "type=AVATAR_BLOCK")
   public String updateAvatarBlock(Model model,
                                   @ModelAttribute("currentBlock") BlockModel blockModel,
-                                  @RequestParam(value = "resetAvatar", required = false, defaultValue = "false") boolean resetAvatar,
                                   @RequestParam(value = "newAvatar", required = false) MultipartFile newAvatar,
+                                  @RequestParam(value = "hasBackground", required = false, defaultValue = "false") boolean hasBackground,
                                   @RequestParam(value = "newBackground", required = false) MultipartFile newBackground,
                                   Authentication authentication) throws Exception {
 
@@ -83,10 +83,9 @@ public class BlockUpdateHandler {
     final Block block = blockRepository.findById(blockModel.getBlockId()).orElseThrow();
     final User user = sessionService.getUser(authentication);
     if (block.props() instanceof AvatarBlockProps avatarProps) {
-      if (resetAvatar) {
-        avatarProps.setAvatarId(AvatarBlockProps.DEFAULT_AVATAR);
-      }
-      if (!newBackground.isEmpty()) {
+      if (!hasBackground) {
+        avatarProps.setBackgroundId(null);
+      } else if (!newBackground.isEmpty()) {
         final byte[] processedImage = imageService.process(newBackground.getInputStream(), imagePresets.avatarBackground());
         final FileSaveRequest request = new FileSaveRequest(user, "background.jpg", new ByteArrayInputStream(processedImage));
         final FileSaveResponse savedBackground = fileService.save(request);
@@ -101,7 +100,7 @@ public class BlockUpdateHandler {
       blockRepository.save(block.withProps(avatarProps));
     }
 
-    rebuildViewModel(model, blockModel);
+    rebuildViewModel(model, blockModel.getBlockId());
     return "private/dashboard";
   }
 
@@ -150,10 +149,12 @@ public class BlockUpdateHandler {
     blockRepository.save(updated);
   }
 
-  private void rebuildViewModel(Model model, BlockModel blockModel) {
-    final Page page = pageRepository.findById(blockModel.getPageId()).orElseThrow();
+  private void rebuildViewModel(Model model, long blockId) {
+    final Block block = blockRepository.findById(blockId).orElseThrow();
+    final Page page = pageRepository.findById(block.pageId()).orElseThrow();
     final Profile profile = profileRepository.findById(page.profileId()).orElseThrow();
 
+    final BlockModel blockModel = blockConverter.convert(block);
     final DashboardModel dashboardModel = dashboardConverter.convert(profile, page);
     model.addAttribute("model", dashboardModel);
     model.addAttribute("currentBlock", blockModel);